feat(animation): fix animations
This fixes the animations and brings in a slightly better word list that
includes what is supposedly an official wordle word list, many of which
aren't in Google's ngrams set, which is kinda weird, but whatever. Word
that aren't in the set get a frequency of 0.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>