fix(ci): ensure unit tests don't watch
Although I don't think this was the case, it appears that it might've
been a situation when `npm run unit:test` would go into watch mode and
not exit. I'm guessing that there's something in `vitest` that checks
for this, but I'd rather be explicit about the tests running once,
rather than coming back and finding that my CI didn't finish because
it was stuck watching the unit tests.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>